feat(require): add support for browserify.require and options.
this is nice because you can require the bundle in your tests just
as you expect folks to in your module.

this is a little weird because requires are not evaluated until you
require them for the first time which breaks the window.__coverage__
test. If you specify require i just test that the code has __coverage__
in it.

couple notes.
- im not sure the exception is super useful
- and when its thrown its really hard to debug because of error propagation.

this also has a fix for console.log within the page. the api .text is
actually a function so now we call it to get the message logged from the
page.